Rename metadata model to config and add corpus type name to the config object
